refactor: card icons location;
