+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Payment Cards Manager — Manage Your Payment Cards Easily</title>
     <meta name="description" content="A React app to manage your payment cards — add, delete, and set default cards." />
-    <link rel="icon" href="./src/assets/images/favicon/credit.png" />
+    <link rel="icon" href="public/credit.png" />
   </head>
   <body>
     <div id="root"></div>
